Refactor to remove src dir and move children
