+++ cmd/accountd/doc.go
