+++ cmd/accountd/integrationtests/testdata/testGetOneUsersSuccess.golden
