+++ cmd/accountd/integrationtests/testdata/testPOSTUserSuccess.golden
