+++ cmd/accountd/integrationtests/testdata/testPUTUserSuccess.golden
