+++ cmd/accountd/internal/config/testdata/..someTestDir/dbpassword
