+++ cmd/accountd/internal/config/testdata/..someTestDirSymLink
