+++ cmd/accountd/internal/config/testdata/dbpassword
