+++ cmd/accountd/main.go
@@ -20,14 +20,14 @@ import (
 	"github.com/juju/errors"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
-	"github.com/youngkin/mockvideo/src/cmd/accountd/handlers"
-	"github.com/youngkin/mockvideo/src/cmd/accountd/handlers/users"
-	"github.com/youngkin/mockvideo/src/cmd/accountd/internal/config"
-	"github.com/youngkin/mockvideo/src/cmd/accountd/usecases"
-	"github.com/youngkin/mockvideo/src/internal/constants"
-	"github.com/youngkin/mockvideo/src/internal/db"
-	user "github.com/youngkin/mockvideo/src/internal/db"
-	"github.com/youngkin/mockvideo/src/internal/logging"
+	"github.com/youngkin/mockvideo/cmd/accountd/handlers"
+	"github.com/youngkin/mockvideo/cmd/accountd/handlers/users"
+	"github.com/youngkin/mockvideo/cmd/accountd/internal/config"
+	"github.com/youngkin/mockvideo/cmd/accountd/usecases"
+	"github.com/youngkin/mockvideo/internal/constants"
+	"github.com/youngkin/mockvideo/internal/db"
+	user "github.com/youngkin/mockvideo/internal/db"
+	"github.com/youngkin/mockvideo/internal/logging"
 
 	log "github.com/sirupsen/logrus"
 )
