+++ cmd/productd/helm/productd/templates/service.yaml
