+++ cmd/productd/helm/productd/templates/tests/test-connection.yaml
